requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ConfigArgParse==1.7
coverage==7.3.1
exceptiongroup==1.1.3
fastapi==0.103.1
Flask==2.3.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
GeoAlchemy2==0.14.1
geojson==3.0.1
gevent==23.9.1
geventhttpclient==2.0.10
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.16.1
MarkupSafe==2.1.3
msgpack==1.0.5
numpy==1.25.2
packaging==23.1
pluggy==1.3.0
psutil==5.9.5
psycopg2==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.0.2
pytest-metadata==3.0.0
python-decouple==3.8
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7
zope.event==5.0
zope.interface==6.0
